Prefer kubernetes informers to hand-rolled API clients
The Kubernetes informer is a well-worn path to implementing
clients that need to be eventually consistent with the k8s
API server. In this case, we're subscribing to the a single
Endpoints object which has the same name as its Kubernetes
Service.

Also, adds tests to verify behavior using a fake Kubernetes client
set.